feat(compiler): extract Escape.ts scalarCount from LiteralForm
